resize fix, trayicon, autostart
