+++ Source/Ruler/RulerForm.resx
